extracted reactive array to reuse in typed handler
